+++ src/components/form/formBlue.jsx
@@ -1,28 +1,6 @@
 import React, { useState } from "react";
 import penform from "../../assets/icons/penForm.svg";
-
-const BoxBlue = ({ titleBox, squareBlue, type, pattern }) => {
-  const [isValid, setIsValid] = useState(true);
-
-  const validateInput = (event) => {
-    setIsValid(event.target.validity.valid);
-  };
-
-  return (
-    <div className="flex flex-col mb-2 w-full px-2">
-      <label htmlFor={titleBox} className="text-LetterColor font-poppinsBold text-sm mb-1">{titleBox}</label>
-      <input
-        type={type}
-        id={titleBox}
-        placeholder={squareBlue}
-        pattern={pattern}
-        required
-        className={`border-2 ${isValid ? 'border-quarteryColor' : 'border-red-500'} rounded-lg py-2 px-4 shadow-xl focus:outline-none bg-white`}
-        onChange={validateInput}
-      />
-    </div>
-  );
-};
+import BoxBlue from "../form/AtomForm";
 
 const FormBlue = () => {
   return (
